pyproject.toml
[tool.poetry]
name = "plus-panner-scheduler"
version = "0.1.0"
description = ""
authors = ["Pedro Gustavo Santana <pedrogustavosantan97@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.111.0"
python-dotenv = "^1.0.1"
invoke = "^2.2.0"
uvicorn = "^0.29.0"
pydantic = "^2.7.1"
aerich = "^0.7.2"
PyJWT = "^2.8.0"
tortoise-orm = "^0.20.1"
websockets = "^12.0"
filelock = "^3.14.0"
requests = "^2.31.0"
asyncpg = "^0.29.0"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
pre-commit = "^3.7.1"


[tool.aerich]
tortoise_orm = "src.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
